+++ hyperparameter_hunter/models.py
@@ -343,20 +343,21 @@ class KerasModel(Model):
         do_predict_proba: Boolean, or int, default=False
             * If False, :meth:`.models.Model.fit` will call :meth:`models.Model.model.predict`
             * If True, it will call :meth:`models.Model.model.predict_proba`, and the values in all
-             columns will be used as the actual prediction values
-            * If `do_predict_proba` is an int, :meth:`.models.Model.fit` will call
-              :meth:`models.Model.model.predict_proba`, as is the case when `do_predict_proba` is
-              True, but the int supplied as `do_predict_proba` declares the column index to use as
-              the actual prediction values
-            * For example, for a model to call the `predict` method, `do_predict_proba=False`
-              (default). For a model to call the `predict_proba` method, and use all of the class
-              probabilities, `do_predict_proba=True`. To call the `predict_proba` method, and use
-              the class probabilities in the first column, `do_predict_proba=0`. To use the second
-              column (index 1) of the result, `do_predict_proba=1` - This often corresponds to the
-              positive class's probabilities in binary classification problems. To use the third
-              column `do_predict_proba=2`, and so on
-            * See the notes for the `do_predict_proba` parameter in the documentation of
-              :class:`environment.Environment` for additional usage notes
+              columns will be used as the actual prediction values
+            * If int, :meth:`.models.Model.fit` will call :meth:`models.Model.model.predict_proba`,
+              as is the case when `do_predict_proba` is True, but the int supplied as
+              `do_predict_proba` declares the column index to use as the actual prediction values
+
+            For example, for a model to call the `predict` method, `do_predict_proba=False`
+            (default). For a model to call the `predict_proba` method, and use all of the class
+            probabilities, `do_predict_proba=True`. To call the `predict_proba` method, and use the
+            class probabilities in the first column, `do_predict_proba=0`. To use the second column
+            (index 1) of the result, `do_predict_proba=1` - This often corresponds to the positive
+            class's probabilities in binary classification problems. To use the third column
+            `do_predict_proba=2`, and so on.
+
+            See the notes for the `do_predict_proba` parameter of
+            :class:`~hyperparameter_hunter.environment.Environment` for additional usage notes
         target_metric: Tuple
             Used by some child classes (like :class:`XGBoostModel`) to provide validation data to
             :meth:`model.fit`